refactor(oa): split obstacle-avoidance orchestration out of behaviors.cpp
